refactor: move widget prop types
