+++ src/index.ts
@@ -1,5 +1,10 @@
 export { default as TurnstileWidget } from './TurnstileWidget.vue'
 export { default as TurnstilePlugin, TurnstileOptionsKey } from './plugin'
 export type { TurnstilePluginOptions } from './plugin'
-export type { Language, LogLevel } from './types.ts'
+export type {
+  Language,
+  LogLevel,
+  TurnstileProps,
+  TurnstileSharedOptions,
+} from './types.ts'
 export type { Theme, WidgetSize, FailureRetryMode, AppearanceMode, RefreshExpiredMode, RefreshTimeoutMode, ExecutionMode } from './turnstile.ts'
