+++ src/types.ts
@@ -72,3 +72,28 @@ export type Language =
 
 export const TESTING_SITEKEY = '1x00000000000000000000AA';
 
+import type { RenderParameters } from './turnstile.ts';
+
+export type TurnstileSharedOptions =
+  | 'action'
+  | 'cData'
+  | 'theme'
+  | 'tabindex'
+  | 'size'
+  | 'retry'
+  | 'retry-interval'
+  | 'appearance'
+  | 'refresh-expired'
+  | 'refresh-timeout'
+  | 'execution'
+  | 'feedback-enabled';
+
+export type TurnstileProps = Pick<RenderParameters, TurnstileSharedOptions> & {
+  /** Every widget has a sitekey. This sitekey is associated with the corresponding widget configuration and is created upon the widget creation. */
+  sitekey?: string;
+  /** Log level for browser logs. Can be debug, info, warn, or error. */
+  logLevel?: LogLevel;
+  /** Language to display. Must be either: auto (default) to use the language that the visitor has chosen, or an ISO 639-1 two-letter language code (e.g. en) or language and country code (e.g. en-US). See https://developers.cloudflare.com/turnstile/reference/supported-languages/ */
+  language?: Language;
+};
+
